Lenovo Tab 3 8 Plus - device_tb8703
Everything needed to compile a stable Version of LineageOS 20 / crDroid for the Lenovo Tab 3 8 Plus 
Including GAPPS

This will not work out of the Box if you are using an older version. You need new Recovery, a zip to partition your device in a new way and need to sign your boot.img.